Switch to new glob api
This commit removes the old Glob implementation and fixes all of the
existing code to work with the new api.
