+++ io/src/main/scala/sbt/io/syntax.scala
@@ -12,9 +12,8 @@ package sbt.io
 
 import java.io.File
 
+import sbt.nio.file.FileGlobApi.FileGlobBuilder
 import sbt.io.PathFinder.Combinator.SingleFilePathFinderCombinator
-import sbt.nio.file.Glob.{ FileGlobBuilder, FileOps }
-import sbt.nio.file.Glob
 
 @deprecated("Alternative is likely to be removed in future versions of sbt", "1.3.0")
 private[sbt] trait Alternative[A, B] {
@@ -33,11 +32,10 @@ sealed trait BaseSyntax extends IOSyntax2 {
       "from File to PathFinder may not be supported in sbt 2"
   )
   def singleFileFinder(file: File): PathFinder = PathFinder(file)
-  implicit def singleFileGlobBuilder(file: File): FileGlobBuilder = new Glob.FileGlobBuilder(file)
+  implicit def singleFileGlobBuilder(file: File): FileGlobBuilder = new FileGlobBuilder(file)
   implicit def singleFilePathFinderCombinator(file: File): PathFinder.Combinator =
     new SingleFilePathFinderCombinator(file)
   implicit def singleFilePathLister(file: File): PathLister = PathLister(file)
-  implicit def singleFileToGlob(file: File): FileOps = new FileOps(file)
 }
 sealed abstract class IOSyntax1 extends BaseSyntax
