+++ io/src/main/scala/sbt/nio/file/package.scala
@@ -1,0 +1,6 @@
+package sbt.nio
+
+package object file {
+  val * = sbt.nio.file.RelativeGlob.*
+  val ** = sbt.nio.file.RelativeGlob.**
+}
