--- io/src/main/scala/sbt/nio/file/proposal/Glob.scala
@@ -1,465 +1,0 @@
-package sbt.nio.file.proposal
-
-import java.io.File
-import java.nio.file.{ FileSystems, Path, Paths }
-import java.util
-
-import sbt.io._
-import sbt.nio.file.proposal.RelativeGlob.{
-  NoPath,
-  PathComponent,
-  SingleComponentMatcher,
-  SingleFileFunctionMatcher,
-  SingleNameFunctionMatcher
-}
-
-import scala.annotation.tailrec
-import scala.collection.JavaConverters._
-import scala.util.Properties
-
-sealed trait Glob {
-
-  /**
-   * Indicates whether a path matches the pattern specified by this [[Glob]].
-   *
-   * @param path the path to match
-   * @return true it the path matches.
-   */
-  def matches(path: Path): Boolean
-}
-
-object Glob {
-  private[this] def comp[T](left: T, right: T)(implicit ordering: Ordering[T]): Int =
-    ordering.compare(left, right)
-  implicit object ordering extends Ordering[Glob] {
-    override def compare(left: Glob, right: Glob): Int = left match {
-      case Pattern(leftRoot, leftRelative) =>
-        right match {
-          case Pattern(rightRoot, rightRelative) =>
-            leftRoot.compareTo(rightRoot) match {
-              case 0 => compare(leftRelative, rightRelative)
-              case i => i
-            }
-          case Root(rightRoot) =>
-            leftRoot.compareTo(rightRoot) match {
-              case 0 => -1
-              case i => i
-            }
-          case _: RelativeGlob => -1
-        }
-      case RelativeGlob(leftMatchers) =>
-        right match {
-          case RelativeGlob(rightMatchers) => comp(leftMatchers, rightMatchers)
-          case _: Pattern                  => 1
-          case _: Root                     => 1
-        }
-      case Root(leftRoot) =>
-        right match {
-          case Root(rightRoot) => leftRoot.compareTo(rightRoot)
-          case _               => -compare(right, left)
-        }
-    }
-  }
-  private object Root {
-    implicit val ordering: Ordering[Root] = Ordering.by(_.root)
-    def apply(value: Path): Root = new Root(value)
-    def unapply(root: Root): Option[Path] = Some(root.root)
-  }
-  private final class Root(val root: Path) extends Glob {
-    require(root.isAbsolute, s"Tried to construct absolute glob from relative path $root")
-    override def matches(path: Path): Boolean = root == path
-    override def toString: String = root.toString
-    override def equals(o: Any): Boolean = o match {
-      case that: Root => this.root == that.root
-      case _          => false
-    }
-    override def hashCode: Int = root.hashCode
-  }
-  private sealed trait Pattern extends Glob {
-    def root: Path
-    def relative: RelativeGlob
-  }
-  private object Pattern {
-    implicit val ordering: Ordering[Pattern] = Ordering.by(p => (p.root, p.relative))
-    def apply(root: Path, relative: RelativeGlob): Pattern = new PatternImpl(root, relative)
-    def unapply(pattern: Pattern): Option[(Path, RelativeGlob)] =
-      Some((pattern.root, pattern.relative))
-  }
-  private final class PatternImpl(override val root: Path, override val relative: RelativeGlob)
-      extends Pattern {
-    override def matches(path: Path): Boolean =
-      path.startsWith(root) && ((path != root) && relative.matches(root.relativize(path)))
-    override def equals(o: Any): Boolean = o match {
-      case that: Pattern =>
-        this.root == that.root && this.relative == that.relative
-      case _ => false
-    }
-    override def hashCode: Int = (root.hashCode * 31) ^ relative.hashCode
-    override def toString: String = s"$root${File.separatorChar}$relative"
-  }
-
-  def apply(file: File): Glob = Root(file.toPath)
-  def apply(file: File, relative: RelativeGlob): Glob = Pattern(file.toPath, relative)
-  def apply(path: Path): Glob = Root(path)
-  def apply(path: Path, relative: RelativeGlob): Glob = Pattern(path, relative)
-  def apply(glob: String): Glob = {
-    val parts = splitter(glob)
-    @tailrec def fullGlob(path: Path, rest: List[String]): Glob = {
-      rest match {
-        case component :: tail if !hasMeta(component) => fullGlob(path.resolve(component), tail)
-        case Nil                                      => Root(path)
-        case _                                        => Pattern(path, RelativeGlob(rest: _*))
-      }
-    }
-    parts match {
-      case h :: rest if !hasMeta(h) =>
-        val base = if (isWin) {
-          if (h.endsWith(":")) Paths.get(h + File.separator) else Paths.get(h)
-        } else {
-          if (h.isEmpty) Paths.get("/") else Paths.get(h)
-        }
-        base match {
-          case p if p.isAbsolute => fullGlob(p, rest)
-          case _                 => RelativeGlob(parts: _*)
-        }
-      case _ => RelativeGlob(parts: _*)
-    }
-  }
-  private[sbt] sealed trait RelativeGlobViewOption
-  private[sbt] object RelativeGlobViewOption {
-    private[sbt] val propName = "sbt.io.implicit.relative.glob.conversion"
-    implicit val default: RelativeGlobViewOption = System.getProperty(propName) match {
-      case "allow"       => Ignore
-      case "error"       => Error
-      case "warn" | null => Warn
-      case p =>
-        val message = s"Unrecognized option $p passed in for $propName. " +
-          "Valid values are: {'allow', 'warn', 'error'}. Setting default to: 'warn'."
-        System.err.println(message)
-        Warn
-    }
-  }
-  private[sbt] case object Warn extends RelativeGlobViewOption
-  private[sbt] case object Error extends RelativeGlobViewOption
-  private[sbt] case object Ignore extends RelativeGlobViewOption
-  private[this] def errorMessage(relative: RelativeGlob, warn: Boolean): String = {
-    val prefix = if (warn) "Warning" else "Error"
-    val action =
-      if (warn) "To disable this warning, "
-      else "To allow implicit conversions using the current jvm working directory, "
-    s"$prefix: Tried to extract the base path for relative glob $relative. $action" +
-      s"re-run the program with java option, -D${RelativeGlobViewOption.propName}=allow"
-  }
-  implicit class GlobOps(val glob: Glob) extends AnyVal {
-    private[sbt] def descendentMatches(path: Path): Boolean = glob match {
-      case Pattern(p, relative) if path.startsWith(p) => relative.matches(p.getFileName)
-    }
-    private[sbt] def toFileFilter: FileFilter = new FileFilter {
-      override def accept(pathname: File): Boolean = glob.matches(pathname.toPath)
-    }
-    private[sbt] def base(implicit option: RelativeGlobViewOption): Path = glob match {
-      case Pattern(root, _) => root
-      case Root(root)       => root
-      case r: RelativeGlob =>
-        option match {
-          case Error  => throw new IllegalArgumentException(errorMessage(r, warn = false))
-          case Ignore => Paths.get("").toAbsolutePath
-          case Warn =>
-            System.err.println(errorMessage(r, warn = true))
-            Paths.get("").toAbsolutePath
-        }
-    }
-    private[sbt] def range: (Int, Int) = glob match {
-      case Pattern(_, relative: RelativeGlob) => RelativeGlob.range(relative)
-      case Root(_)                            => (0, 0)
-      case relative: RelativeGlob             => RelativeGlob.range(relative)
-    }
-    def /(that: RelativeGlob): Glob = glob match {
-      case Pattern(root, relative) => Pattern(root, relative / that)
-      case Root(path) =>
-        that match {
-          case PathComponent(component) => Root(path.resolve(component))
-          case t                        => Pattern(path, t)
-        }
-      case r: RelativeGlob => r / that
-    }
-  }
-  implicit def stringToGlob(glob: String): Glob = Glob(glob)
-  final class PathOps(val path: Path) extends AnyVal {
-    def glob: Glob = Root(path)
-    def /(component: String): Path = path.resolve(component)
-    def /(relative: RelativeGlob): Glob = Pattern(path, relative)
-
-  }
-  final class FileOps(val file: File) extends AnyVal {
-    def glob: Glob = Root(file.toPath)
-    def /(relative: RelativeGlob): Glob = Pattern(file.toPath, relative)
-  }
-  private[this] val windowsEscapable = "(){}"
-  private[this] val allMeta = "*{([?"
-  private[file] val hasMeta: String => Boolean = _.exists(allMeta.contains(_))
-  private[this] val isWin = Properties.isWin
-  private[this] val splitter: String => List[String] = {
-    if (isWin) { glob =>
-      {
-        val stringBuilder = new StringBuilder(glob.length)
-        val components = new util.ArrayList[String]
-        val array = glob.toCharArray
-        val separator = File.separatorChar
-        @tailrec def fillComponents(index: Int): Unit = index match {
-          case i if i < array.length =>
-            array(i) match {
-              case `separator` =>
-                val nextIndex = i + 1
-                if (nextIndex < array.length) {
-                  array(nextIndex) match {
-                    case c if windowsEscapable.contains(c) =>
-                      stringBuilder.append(separator).append(c)
-                    case c =>
-                      components.add(stringBuilder.toString)
-                      stringBuilder.clear
-                      stringBuilder.append(c)
-                  }
-                  fillComponents(nextIndex + 1)
-                } else {
-                  components.add(stringBuilder.toString)
-                  ()
-                }
-              case c =>
-                stringBuilder.append(c)
-                fillComponents(i + 1)
-            }
-          case _ =>
-            components.add(stringBuilder.toString)
-            ()
-        }
-        fillComponents(index = 0)
-        components.asScala.toList
-      }
-    } else {
-      _.split(File.separatorChar).toList
-    }
-  }
-}
-
-sealed trait RelativeGlob extends Glob {
-  private[file] def matchers: List[RelativeGlob.Matcher]
-  def /(that: RelativeGlob): RelativeGlob = RelativeGlob(this.matchers ::: that.matchers)
-}
-case object RecursiveGlob extends SingleComponentMatcher with RelativeGlob {
-  override def glob: String = "**"
-  def matches(path: Path): Boolean = true
-}
-case object AnyPath extends SingleComponentMatcher with RelativeGlob {
-  override def glob: String = "*"
-  override def matches(path: Path): Boolean = path.getNameCount == 1
-}
-object RelativeGlob {
-  val ** = RecursiveGlob
-  val * = AnyPath
-  def apply(matchers: String*): RelativeGlob =
-    new RelativeGlobImpl(matchers.map(Matcher.apply).toList)
-  private[sbt] def apply(matchers: List[Matcher]): RelativeGlob = new RelativeGlobImpl(matchers)
-  private[sbt] def unapply(relativeGlob: RelativeGlob): Option[List[Matcher]] =
-    Some(relativeGlob.matchers)
-  implicit def stringToRelativeGlob(glob: String): RelativeGlob = Glob(glob) match {
-    case r: RelativeGlob => r
-    case _ =>
-      throw new IllegalArgumentException(
-        s"Tried to construct a relative glob using an absolute path $glob")
-  }
-  implicit val ordering: Ordering[RelativeGlob] = Ordering.by(_.matchers)
-  private[file] def range(relative: RelativeGlob): (Int, Int) = {
-    val res = relative.matchers.foldLeft((0, 0)) {
-      case ((0, 0), RecursiveGlob)                => (1, Int.MaxValue)
-      case (r @ (_, Int.MaxValue), RecursiveGlob) => r
-      case ((min, Int.MaxValue), _)               => (min + 1, Int.MaxValue)
-      case ((min, max), _)                        => (min + 1, max + 1)
-    }
-    res
-  }
-
-  private final class RelativeGlobImpl(val matchers: List[Matcher]) extends RelativeGlob {
-    override def matches(path: Path): Boolean = {
-      val count = path.getNameCount
-      @tailrec def impl(currentIndex: Int, matchers: List[Matcher]): Boolean = matchers match {
-        case RecursiveGlob :: Nil => count > 0
-        case RecursiveGlob :: matchersTail =>
-          currentIndex match {
-            case i if i < count => recursiveMatches(matchersTail, i)
-            case _              => false
-          }
-        case m :: Nil if currentIndex == count - 1 => m.matches(path.getFileName)
-        case m :: matchersTail =>
-          currentIndex match {
-            case i if i < count && m.matches(path.getName(i)) => impl(i + 1, matchersTail)
-            case _                                            => false
-          }
-        case Nil => currentIndex == count
-      }
-      def recursiveMatches(remaining: List[Matcher], currentIndex: Int): Boolean = {
-        remaining match {
-          case Nil                => true
-          case nameMatcher :: Nil => nameMatcher.matches(path.getFileName)
-          case _ =>
-            @tailrec def recursiveImpl(index: Int): Boolean = index match {
-              case i if i < count => impl(i, remaining) || recursiveImpl(i + 1)
-              case _              => false
-            }
-            recursiveImpl(currentIndex)
-        }
-      }
-      impl(currentIndex = 0, matchers)
-    }
-    override def hashCode: Int = matchers.hashCode
-    override def toString: String = matchers.mkString(File.separator)
-    override def equals(o: Any): Boolean = o match {
-      case that: RelativeGlob => this.matchers == that.matchers
-      case _                  => false
-    }
-  }
-
-  private[file] sealed trait Matcher extends RelativeGlob
-  private[file] object Matcher {
-    implicit object ordering extends Ordering[Matcher] {
-      override def compare(x: Matcher, y: Matcher): Int = x match {
-        case RecursiveGlob =>
-          y match {
-            case RecursiveGlob => 0
-            case _             => -1
-          }
-        case AnyPath =>
-          y match {
-            case AnyPath       => 0
-            case RecursiveGlob => 1
-            case _             => -1
-          }
-        case spm: SingleComponentMatcher =>
-          y match {
-            case that: SingleComponentMatcher => spm.glob.compareTo(that.glob)
-            case _                            => 1
-          }
-        case _: SingleNameFunctionMatcher =>
-          y match {
-            case _: SingleNameFunctionMatcher => 0
-            case _                            => 1
-          }
-        case _: SingleFileFunctionMatcher =>
-          y match {
-            case _: SingleFileFunctionMatcher => 0
-            case _                            => 1
-          }
-      }
-    }
-    implicit val listOrdering: Ordering[List[Matcher]] = new Ordering[List[Matcher]] {
-      override def compare(left: List[Matcher], right: List[Matcher]): Int = {
-        val leftIt = left.iterator
-        val rightIt = right.iterator
-        while (leftIt.hasNext && rightIt.hasNext) {
-          val res = ordering.compare(leftIt.next, rightIt.next)
-          if (res != 0) return res
-        }
-        Ordering.Boolean.compare(leftIt.hasNext, rightIt.hasNext)
-      }
-    }
-    def apply(glob: String): Matcher = glob match {
-      case "**"                  => RecursiveGlob
-      case "*"                   => AnyPath
-      case g if !Glob.hasMeta(g) => PathComponent(g)
-      case g                     => new GlobMatcher(g)
-    }
-  }
-  private[file] case object NoPath extends SingleComponentMatcher with RelativeGlob {
-    override def glob: String = "<null>"
-    override def matches(path: Path): Boolean = false
-  }
-  private[file] class SingleNameFunctionMatcher(private val f: String => Boolean) extends Matcher {
-    override private[sbt] def matchers: List[Matcher] = this :: Nil
-    override def matches(path: Path): Boolean = f(path.getFileName.toString)
-    override def equals(o: Any): Boolean = o match {
-      case that: SingleNameFunctionMatcher => this.f == that.f
-      case _                               => false
-    }
-    override def hashCode: Int = f.hashCode
-    override def toString: String = f.toString
-  }
-  private[file] class SingleFileFunctionMatcher(private val base: Path,
-                                                private val fileFilter: FileFilter)
-      extends Matcher {
-    override def matches(path: Path): Boolean = fileFilter.accept(base.resolve(path).toFile)
-    override private[sbt] def matchers: List[Matcher] = this :: Nil
-    override def equals(o: Any): Boolean = o match {
-      case that: SingleFileFunctionMatcher =>
-        this.base == that.base && this.fileFilter == that.fileFilter
-      case _ => false
-    }
-    override def hashCode: Int = (base.hashCode * 31) ^ fileFilter.hashCode
-    override def toString: String = fileFilter.toString
-  }
-  private[file] sealed trait SingleComponentMatcher extends Matcher {
-    def glob: String
-    override final val matchers: List[Matcher] = this :: Nil
-    override def toString: String = glob
-  }
-  private[file] object PathComponent {
-    def apply(component: String): PathComponent = new PathComponent(component)
-    def unapply(pathComponent: PathComponent): Option[String] = Some(pathComponent.glob)
-  }
-  private[file] final class PathComponent private (override val glob: String)
-      extends SingleComponentMatcher {
-    override def matches(path: Path): Boolean =
-      path.getNameCount == 1 && path.getFileName.toString == glob
-    override def equals(o: Any): Boolean = o match {
-      case that: PathComponent => this.glob == that.glob
-      case _                   => false
-    }
-    override def hashCode: Int = glob.hashCode
-  }
-  private final class GlobMatcher(override val glob: String) extends SingleComponentMatcher {
-    private[this] val (prefix, pattern) = glob.indexOf(":") match {
-      case -1 => ("glob", glob)
-      case i  => (glob.substring(0, i), glob.substring(i + 1))
-    }
-    private[this] val matcher = FileSystems.getDefault.getPathMatcher(s"$prefix:$pattern")
-    private[this] val excludeHiddenFiles = prefix == "glob" && pattern.startsWith("*")
-    override def matches(path: Path): Boolean =
-      matcher.matches(path) && !(excludeHiddenFiles && path.getFileName.toString.startsWith("."))
-    override def equals(o: Any): Boolean = o match {
-      case that: GlobMatcher => this.glob == that.glob
-      case _                 => false
-    }
-    override def hashCode(): Int = glob.hashCode
-  }
-}
-
-private[sbt] object FileGlobApi {
-  private[sbt] final class FileGlobBuilder(val file: File) extends AnyVal {
-    def /(component: String): Glob = Glob(file.toPath.resolve(component))
-    def \(component: String): Glob = this / component
-    def glob: Glob = Glob(file)
-    def glob(filter: FileFilter): Glob = filter match {
-      case AllPassFilter => Glob(file.toPath, AnyPath)
-      case f             => Glob(file.toPath, convert(f))
-    }
-    def *(filter: FileFilter): Glob = glob(filter)
-    def globRecursive(filter: FileFilter): Glob = filter match {
-      case AllPassFilter => Glob(file.toPath, RecursiveGlob)
-      case f             => Glob(file.toPath, RecursiveGlob / convert(f))
-    }
-    def allPaths: Glob = globRecursive(sbt.io.AllPassFilter)
-    def **(filter: FileFilter): Glob = globRecursive(filter)
-
-    private def convert(filter: FileFilter): RelativeGlob = filter match {
-      case AllPassFilter => AnyPath
-      case NothingFilter => NoPath
-      case ef: ExtensionFilter =>
-        ef.extensions match {
-          case Seq(extension) => RelativeGlob(s"*.$extension")
-          case extensions     => RelativeGlob(s"*.${extensions.mkString("{", ",", "}")}")
-        }
-      case exactFilter: ExactFilter => RelativeGlob(exactFilter.matchName)
-      case nf: NameFilter           => new SingleNameFunctionMatcher(nf.accept(_: String))
-      case f: FileFilter            => new SingleFileFunctionMatcher(file.toPath, f)
-    }
-  }
-}
