+++ io/src/test/scala/sbt/internal/io/WatchServiceBackedObservableSpec.scala
@@ -8,8 +8,8 @@ import java.util.concurrent.{ ConcurrentHashMap, CountDownLatch, TimeUnit }
 import org.scalatest.FlatSpec
 import sbt.internal.nio.WatchServiceBackedObservable
 import sbt.io._
-import sbt.nio.file.Glob
-import sbt.nio.syntax._
+import sbt.nio.file._
+import sbt.nio.file.syntax._
 
 import scala.collection.JavaConverters._
 import scala.concurrent.duration._
@@ -33,7 +33,7 @@ class WatchServiceBackedObservableSpec extends FlatSpec {
       val latch = new CountDownLatch(1)
       val file = subdir.resolve("file")
       observable.addObserver(e => if (e.path == file) latch.countDown())
-      observable.register(path / "**")
+      observable.register(path.toGlob / RecursiveGlob)
       Files.createFile(file)
       assert(latch.await(1, TimeUnit.SECONDS))
     } finally observable.close()
